whatsapp/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
twilio>=8.10.0
python-multipart>=0.0.6
pydantic>=2.7.4
python-dotenv>=1.0.0
pymongo>=4.0.0
openai>=1.0.0
tiktoken>=0.4.0
supabase>=2.0.0